Move report_date_time() to external.utils.functions.
