+++ components/server/src/external/routes/measurement.py
@@ -20,8 +20,9 @@ from external.database import sessions
 from model.data import SourceData
 from model.measurement import Measurement
 from model.metric import Metric
-from server_utilities.functions import report_date_time
 from server_utilities.type import MetricId, SourceId
+
+from ..utils.functions import report_date_time
 
 from .plugins.auth_plugin import EDIT_ENTITY_PERMISSION
