+++ components/server/src/external/routes/report.py
@@ -24,12 +24,11 @@ from server_utilities.functions import (
     DecryptionError,
     check_url_availability,
     iso_timestamp,
-    report_date_time,
     uuid,
 )
 from server_utilities.type import ReportId
 
-from ..utils.functions import sanitize_html
+from ..utils.functions import report_date_time, sanitize_html
 
 from .plugins.auth_plugin import EDIT_REPORT_PERMISSION
