+++ components/server/src/external/utils/functions.py
@@ -2,10 +2,23 @@
 
 import re
 
+import bottle
+
 # Bandit complains that "Using autolink_html to parse untrusted XML data is known to be vulnerable to XML attacks",
 # and Dlint complains 'insecure use of XML modules, prefer "defusedxml"'
 # but we give autolink_html clean html, so ignore the warning:
 from lxml.html.clean import autolink_html, clean_html  # noqa: DUO107, # nosec, pylint: disable=no-name-in-module
+
+from server_utilities.functions import iso_timestamp
+
+
+def report_date_time() -> str:
+    """Return the report date requested as query parameter if it's in the past, else return an empty string."""
+    if report_date_string := dict(bottle.request.query).get("report_date"):
+        iso_report_date_string = str(report_date_string).replace("Z", "+00:00")
+        if iso_report_date_string < iso_timestamp():
+            return iso_report_date_string
+    return ""
 
 
 def sanitize_html(html_text: str) -> str:
